requirements.txt
torch==2.2.2 
torchvision==0.17.2 
torchaudio==2.2.2
git+https://github.com/openai/whisper.git
transformers==4.43.2
streamlit==1.37.0
keybert==0.8.4
paddlepaddle==2.6.1
paddleocr==2.7.3
punctuators==0.0.7
numpy==1.26.4
tqdm==4.66.4
langchain==0.0.248
pandas==2.2.2
opencv-python
llama-cpp-python
graphviz
pathvalidate